### 2024-12-10 #7577 @editorialbot
[REVIEW]: Af2-analysis: a Python package for Alphafold 2 analysis
**Submitting author:** @samuelmurail (<a href="http://orcid.org/0000-0002-3842-5333">Samuel Murail</a>)
**Repository:** https://github.com/samuelmurail/af_analysis
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.4
