+++ src/af_analysis/analysis.py
@@ -32,6 +32,27 @@ def get_pae(json_file):
 
     if json_file is None:
         return None
+    
+    if json_file.endswith(".json"):
+        return extract_pae_json(json_file)
+    elif json_file.endswith(".npz"):
+        return extract_pae_npz(json_file)
+    else:
+        raise ValueError("Unknown file format.")
+
+def extract_pae_json(json_file):
+    """Get the PAE matrix from a json file.
+
+    Parameters
+    ----------
+    json_file : str
+        Path to the json file.
+    
+    Returns
+    -------
+    np.array
+        PAE matrix.
+    """
 
     with open(json_file) as f:
         local_json = json.load(f)
@@ -45,6 +66,25 @@ def get_pae(json_file):
 
     return pae_array
 
+
+def extract_pae_npz(npz_file):
+    """Get the PAE matrix from a json file.
+
+    Parameters
+    ----------
+    npz_file : str
+        Path to the npz file.
+    
+    Returns
+    -------
+    np.array
+        PAE matrix.
+    """
+
+    data_npz = np.load(npz_file)
+    pae_array = data_npz['pae']
+
+    return pae_array
 
 def extract_fields_json(json_file, fields):
     """Get the PAE matrix from a json file.
@@ -223,6 +263,9 @@ def pdockq2(data, verbose=True):
 
     disable = False if verbose else True
 
+    if 'json' not in data.df.columns:
+        raise ValueError("No json column found in the dataframe. pae scores are required to compute pdockq2.")
+
     for pdb, json_path in tqdm(
         zip(data.df["pdb"], data.df["json"]), total=len(data.df["pdb"]), disable=disable
     ):
@@ -274,6 +317,9 @@ def inter_chain_pae(data, fun=np.mean, verbose=True):
     pae_list = []
 
     disable = False if verbose else True
+
+    if 'json' not in data.df.columns:
+        raise ValueError("No json column found in the dataframe. pae scores are required to compute pdockq2.")
 
     for query, json_path in tqdm(
         zip(data.df["query"], data.df["json"]),
