### 2024-12-12 362480e
Rename column to be consistent with colabfold
