### 2024-12-12 57f7aaf
Update test and uniformise AF3 column
