+++ src/af_analysis/format/boltz1.py
@@ -72,11 +72,13 @@ def read_dir(directory):
     log_pd = pd.DataFrame(log_dict_list)
 
     # Update column names
-    log_pd = log_pd.rename(columns={
-        'confidence_score': 'ranking_confidence',
-        'ptm': 'pTM',
-        'iptm': 'ipTM',
-    })
+    log_pd = log_pd.rename(
+        columns={
+            "confidence_score": "ranking_confidence",
+            "ptm": "pTM",
+            "iptm": "ipTM",
+        }
+    )
 
     # To ensure that tests are consistent across different systems
     # we sort the dataframe by pdb
