### 2024-12-12 6ba8d2f
Fix a misunderstanding concerning ranking_confidence
