+++ src/af_analysis/format/boltz1.py
@@ -74,11 +74,13 @@ def read_dir(directory):
     # Update column names
     log_pd = log_pd.rename(
         columns={
-            "confidence_score": "ranking_confidence",
+            # https://github.com/jwohlwend/boltz/issues/73
+            #"confidence_score": "ranking_confidence",
             "ptm": "pTM",
             "iptm": "ipTM",
         }
     )
+    log_pd["ranking_confidence"] = 0.2*log_pd["pTM"] + 0.8*log_pd["ipTM"]
 
     # To ensure that tests are consistent across different systems
     # we sort the dataframe by pdb
