### 2024-12-13 b15f64c
Add analysis for boltz1
