+++ requirements.txt
@@ -1,4 +1,4 @@
-pdb_numpy>=0.0.11
+pdb_numpy>=0.0.12
 pandas>=1.3.4
 numpy>=1.21
 tqdm>=4.0
