### 2024-12-15 272c82c
Fix a warning with pandas
