### 2024-12-16 #7577 @rvhonorato
Two coments about the documentation,
- **Example usage**: An `example.ipynb` is provided but it's not mentioned in the README. I think having an example notebook is great, would be better if it was more obvious for the user on how to find it. It also has some typos `af2_analysis` whereas it should be `af_analysis`. This is also in the `.gif` provided in the plto section.
- **Functionality documentation**: the user documentation is quite good, there doesn't seem to be any code documenation but I don't think this is an issue for this sort of application.
