### 2024-12-16 61d7fc1
Update logo and try to fix nglview html.
