+++ docs/paper/paper.bib
@@ -124,3 +124,54 @@
   doi = {10.1038/s41467-022-28865-w},
   publisher={Nature Publishing Group UK London}
 }
+
+
+@article{abramson_accurate_2024,
+	title = {Accurate structure prediction of biomolecular interactions with {AlphaFold} 3},
+	volume = {630},
+	rights = {2024 The Author(s)},
+	issn = {1476-4687},
+	url = {https://www.nature.com/articles/s41586-024-07487-w},
+	doi = {10.1038/s41586-024-07487-w},
+	abstract = {The introduction of {AlphaFold} 21 has spurred a revolution in modelling the structure of proteins and their interactions, enabling a huge range of applications in protein modelling and design2–6. Here we describe our {AlphaFold} 3 model with a substantially updated diffusion-based architecture that is capable of predicting the joint structure of complexes including proteins, nucleic acids, small molecules, ions and modified residues. The new {AlphaFold} model demonstrates substantially improved accuracy over many previous specialized tools: far greater accuracy for protein–ligand interactions compared with state-of-the-art docking tools, much higher accuracy for protein–nucleic acid interactions compared with nucleic-acid-specific predictors and substantially higher antibody–antigen prediction accuracy compared with {AlphaFold}-Multimer v.2.37,8. Together, these results show that high-accuracy modelling across biomolecular space is possible within a single unified deep-learning framework.},
+	pages = {493--500},
+	number = {8016},
+	journaltitle = {Nature},
+	author = {Abramson, Josh and Adler, Jonas and Dunger, Jack and Evans, Richard and Green, Tim and Pritzel, Alexander and Ronneberger, Olaf and Willmore, Lindsay and Ballard, Andrew J. and Bambrick, Joshua and Bodenstein, Sebastian W. and Evans, David A. and Hung, Chia-Chun and O’Neill, Michael and Reiman, David and Tunyasuvunakool, Kathryn and Wu, Zachary and Žemgulytė, Akvilė and Arvaniti, Eirini and Beattie, Charles and Bertolli, Ottavia and Bridgland, Alex and Cherepanov, Alexey and Congreve, Miles and Cowen-Rivers, Alexander I. and Cowie, Andrew and Figurnov, Michael and Fuchs, Fabian B. and Gladman, Hannah and Jain, Rishub and Khan, Yousuf A. and Low, Caroline M. R. and Perlin, Kuba and Potapenko, Anna and Savy, Pascal and Singh, Sukhdeep and Stecula, Adrian and Thillaisundaram, Ashok and Tong, Catherine and Yakneen, Sergei and Zhong, Ellen D. and Zielinski, Michal and Žídek, Augustin and Bapst, Victor and Kohli, Pushmeet and Jaderberg, Max and Hassabis, Demis and Jumper, John M.},
+	urldate = {2024-06-28},
+	date = {2024-06},
+	langid = {english},
+	note = {Publisher: Nature Publishing Group},
+	keywords = {Structural biology, Drug discovery, Machine learning, Protein structure predictions},
+}
+
+@misc{wohlwend_boltz-1_2024,
+	title = {Boltz-1 Democratizing Biomolecular Interaction Modeling},
+	rights = {© 2024, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), {CC} {BY} 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
+	url = {https://www.biorxiv.org/content/10.1101/2024.11.19.624167v1},
+	doi = {10.1101/2024.11.19.624167},
[…]
