+++ docs/source/analysis.rst
@@ -48,14 +48,18 @@ Scores
 pDockq
 ------
 
-pDockQ (Predicted DockQ) [#pdockq]_ is a metric that predicts the quality of protein-protein interactions in dimer models.  It uses a sigmoid function that considers the number of contacts in the interface (:math:`log(number \: of \: interface \: contacts)`) and the average pLDDT (predicted lDDT score) of the interface residues (:math:`\overline{plDDT_{interface}}`).
+pDockQ (Predicted DockQ) [#pdockq]_ is a metric that predicts the quality of protein-protein interactions in dimer models.
+It uses a sigmoid function that considers the number of contacts in the interface
+(:math:`log(number \: of \: interface \: contacts)`) and the average pLDDT (predicted lDDT score) of the
+interface residues (:math:`\overline{plDDT_{interface}}`).
 
 
 
 Implementation was inspired from `https://gitlab.com/ElofssonLab/FoldDock/-/blob/main/src/pdockq.py <https://gitlab.com/ElofssonLab/FoldDock/-/blob/main/src/pdockq.py>`_.
 
 
-The ``pdockq()`` function calculates the pDockq [#pdockq]_ score for each model in the dataframe. The pDockq score ranges from 0 to 1, with higher scores indicating better model quality.
+The ``pdockq()`` function calculates the pDockq [#pdockq]_ score for each model in the dataframe.
+The pDockq score ranges from 0 to 1, with higher scores indicating better model quality.
 
 .. math::
     pDockQ = \frac{L}{1 + e^{-k (x-x_{0})}} + b
@@ -89,7 +93,11 @@ For the *multiple* pDockQ or `mpDockQ` [#mpdockq]_ this values are used:
 pDockq2
 -------
 
-pDockQ2, or Predicted DockQ version 2 [#pdockq2]_, is a metric used to estimate the quality of individual interfaces in multimeric protein complex models. Unlike the original pDockQ, pDockQ2 incorporates AlphaFold-Multimer's Predicted Aligned Error (PAE) in its calculation, making it more sensitive to large, incorrect interfaces that might have high confidence scores based solely on interface size and pLDDT. pDockQ2 scores range from 0 to 1, with higher scores indicating better interface quality.
+pDockQ2, or Predicted DockQ version 2 [#pdockq2]_, is a metric used to estimate the quality of individual interfaces in
+multimeric protein complex models. Unlike the original pDockQ, pDockQ2 incorporates AlphaFold-Multimer's Predicted
+Aligned Error (PAE) in its calculation, making it more sensitive to large, incorrect interfaces that might have high
+confidence scores based solely on interface size and pLDDT. pDockQ2 scores range from 0 to 1, with higher scores
+indicating better interface quality.
 
 .. math::
     pDockQ_2 = \frac{L}{1 + exp [-k*(X_i-X_0)]} + b
@@ -114,14 +122,25 @@ Implementation was inspired from `https://gitlab.com/ElofssonLab/afm-benchmark/-
 LIS Score
 ---------
 
-The Local Interaction Score (LIS) [#LIS]_ is a metric specifically designed to predict the likelihood of direct protein-protein interactions (PPIs) using output data from AlphaFold-Multimer [#AF2M].
-Unlike metrics like interface pTM (ipTM), which measures the overall structural accuracy of a predicted complex, LIS focuses on areas within the predicted interface that have low Predicted Aligned Error (PAE) values. These low PAE values, often visualized as blue regions in AlphaFold output maps, represent areas of high confidence in the interaction prediction
+The Local Interaction Score (LIS) [#LIS]_ is a metric specifically designed to predict the likelihood of direct
[…]
