### 2024-12-18 9c1ec6c
Add link to google colab notebook
