### 2024-12-18 #7577 @samuelmurail
Dear @rvhonorato,

Thank you for your thoughtful feedback and suggestions. I have now addressed the points raised:

> Two coments about the documentation,
>
> * **Example usage**: An `example.ipynb` is provided but it's not mentioned in the README. I think having an example notebook is great, would be better if it was more obvious for the user on how to find it. It also has some typos `af2_analysis` whereas it should be `af_analysis`. This is also in the `.gif` provided in the plto section.

I’ve updated the `README` to clearly mention the `basic_example.ipynb` notebook (The example which was already present in the Documentation web site). I also added a google colab notebook to allow user to test the library:
[![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/samuelmurail/af_analysis/blob/main/basic_example_colab.ipynb)

> * **Functionality documentation**: the user documentation is quite good, there doesn't seem to be any code documenation but I don't think this is an issue for this sort of application.

I agree the code documentation is not complete, I preferred to explain it in the user documentation.

> One small comment about the paper:
>
> * **State of the field**: I see that it was submitted before the release of AlphaFold3, I think it might be worth it - for visibility and discoverability purposes for the paper to be update to also account for AF3.

Yes, you are right, the release of the AF3 code has had a bit of an impact. I have taken AF3, Chai-1 and Boltz-1 into account. I have also added support in the library for analysing Boltz-1 models.
