### 2024-12-18 fa8c7e3
Add chai1 test set
