### 2024-12-18 4e0e208
Add chai-1 support
