+++ src/af_analysis/test/test_format.py
@@ -231,3 +231,42 @@ def test_boltz1():
     np.testing.assert_allclose(
         np.array(list(my_data.df["ipTM"])), np.array(expected_iptm), atol=precision
     )
+
+
+def test_chai1():
+    data_path = os.path.join(TEST_FILE_PATH, "chai1_prot_dna_ligand")
+
+    my_data = af_analysis.Data(data_path)
+
+    assert my_data.format == "chai1"
+    assert len(my_data.df) == 5
+    print(my_data.df.columns)
+    assert (
+        my_data.df.columns
+        == np.array(
+            [
+                'pdb', 'model', 'ranking_confidence', 'pTM', 'ipTM', 'per_chain_ptm',
+       'per_chain_pair_iptm', 'has_inter_chain_clashes', 'chain_chain_clashes',
+       'query', 'format'
+            ]
+        )
+    ).all()
+
+    query = my_data.df.iloc[0]["query"]
+
+    assert my_data.chain_length[query] == [586, 26, 13, 13, 1]
+    assert my_data.chains[query] == ["A", "B", "C", "D", "E"]
+
+    # There should be 0 relaxed structures
+
+    assert "relaxed_pdb" not in my_data.df.columns
+    print(my_data.df.iloc[:, :])
+    assert list(my_data.df["model"]) == list(range(5))
+    print(my_data.df["ipTM"])
+    expected_iptm = [ 0.954016, 0.953741, 0.953701, 0.953696, 0.953734]
+
+    precision = 0.01
+
+    np.testing.assert_allclose(
+        np.array(list(my_data.df["ipTM"])), np.array(expected_iptm), atol=precision
+    )
