+++ README.md
@@ -22,6 +22,7 @@ generated by:
 * [AlphaFold-Multimer][AF2-M]
 * [AlphaPulldown][AlphaPulldown]
 * [Boltz1][Boltz1].
+* [Chai-1][Chai1].
 
 
 Source code repository:
@@ -184,6 +185,7 @@ This project is licensed under the GNU General Public License version 2 - see th
 - Kim et al. bioRxiv (2024) doi: [10.1101/2024.02.19.580970][LIS]
 - Yu et al. Bioinformatics (2023) doi: [10.1093/bioinformatics/btac749][AlphaPulldown]
 - Wohlwend et al. bioRxiv (2024) doi: [10.1101/2024.11.19.624167][Boltz1]
+- Chai Discovery et al. bioRxiv (2024) doi:[10.1101/2024.10.10.615955v2][Chai1]
 
 
 [AF2]: https://www.nature.com/articles/s41586-021-03819-2 "Jumper et al. Nature (2021) doi: 10.1038/s41586-021-03819-2"
@@ -195,3 +197,4 @@ This project is licensed under the GNU General Public License version 2 - see th
 [LIS]: https://www.biorxiv.org/content/10.1101/2024.02.19.580970v1 "Kim et al. bioRxiv (2024) doi: 10.1101/2024.02.19.580970 "
 [AlphaPulldown]: https://doi.org/10.1093/bioinformatics/btac749 "Yu et al. Bioinformatics (2023) doi: 10.1093/bioinformatics/btac749"
 [Boltz1]: https://doi.org/10.1101/2024.11.19.624167 "Wohlwend et al. bioRxiv (2024) doi: 10.1101/2024.11.19.624167"
+[Chai1]: https://doi.org/10.1101/2024.10.10.615955v2 "Chai Discovery et al. bioRxiv (2024) doi: 10.1101/2024.10.10.615955v2"
