### 2025-01-07 #7577 @AnjaConev
Very nice and relevant work. The package provides very useful functionalities for the community.
A few minor comments:

- Automated tests: pytest test_clustering.py is failing for me. I use Python version 3.11.7 and Rocky Linux 9.3 OS.
![image](https://github.com/user-attachments/assets/ec7fceb7-51cf-4f2e-82ad-ce3e37ab9bc4)
clustering works when running it outside of the tests though.

- State of the field: You could add a paragraph to the paper about similar packages in the field and how yours is different.
