### 2025-01-08 8815ac3
Update doc and paper
