+++ docs/paper/paper.bib
@@ -175,3 +175,58 @@ Section: New Results},
 	note = {Pages: 2024.10.10.615955
 Section: New Results},
 }
+
+@article{raouraoua_massivefold_2024,
+	title = {{MassiveFold}: unveiling {AlphaFold}’s hidden potential with optimized and parallelized massive sampling},
+	volume = {4},
+	copyright = {2024 The Author(s)},
+	issn = {2662-8457},
+	shorttitle = {{MassiveFold}},
+	url = {https://www.nature.com/articles/s43588-024-00714-4},
+	doi = {10.1038/s43588-024-00714-4},
+	abstract = {Massive sampling in AlphaFold enables access to increased structural diversity. In combination with its efficient confidence ranking, this unlocks elevated modeling capabilities for monomeric structures and foremost for protein assemblies. However, the approach struggles with GPU cost and data storage. Here we introduce MassiveFold, an optimized and customizable version of AlphaFold that runs predictions in parallel, reducing the computing time from several months to hours. MassiveFold is scalable and able to run on anything from a single computer to a large GPU infrastructure, where it can fully benefit from all the computing nodes.},
+	language = {en},
+	number = {11},
+	urldate = {2025-01-08},
+	journal = {Nat Comput Sci},
+	author = {Raouraoua, Nessim and Mirabello, Claudio and Véry, Thibaut and Blanchet, Christophe and Wallner, Björn and Lensink, Marc F. and Brysbaert, Guillaume},
+	month = nov,
+	year = {2024},
+	note = {Publisher: Nature Publishing Group},
+	keywords = {Machine learning, Protein structure predictions, Software},
+	pages = {824--828},
+}
+
+@misc{alvarez-salmoral_alphabridge_2024,
+	title = {{AlphaBridge}: tools for the analysis of predicted macromolecular complexes},
+	copyright = {© 2024, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution-NonCommercial-NoDerivs 4.0 International), CC BY-NC-ND 4.0, as described at http://creativecommons.org/licenses/by-nc-nd/4.0/},
+	shorttitle = {{AlphaBridge}},
+	url = {https://www.biorxiv.org/content/10.1101/2024.10.23.619601v1},
+	doi = {10.1101/2024.10.23.619601},
+	abstract = {Artificial intelligence (AI)-powered protein structure prediction methods have revolutionised how life scientists explore macromolecular function. Using AI to predict the structure of macromolecular complexes, is gaining attention for modelling known interactions and evaluating the likelihood of proteins forming multimers or interacting with other proteins or nucleic acids. There is a growing need for tools to efficiently evaluate these predicted models. We introduce new tools that use AlphaFold3’s “predicted local-distance difference test” (pLDDT), “predicted aligned error” (PAE), and “predicted distance error” (PDE) matrices in a graph-based community clustering approach to label sequence motifs involved in binary interactions. The resulting “interaction image” is processed through a multidimensional image algorithm to cluster interacting sequence motifs in three-dimensional binary interfaces. This method allows us to present the interaction information between multiple proteins and nucleic acids back to two-dimensional space using chord diagrams. These “AlphaBridge” diagrams summarise predicted interfaces and intermolecular interactions, including prediction confidence and sequence conservation scores. They are valuable for efficient screening of predictions, ranking, and scoring the confidence of predicted interactions, prior to more detailed (and resource intensive) analysis.},
+	language = {en},
+	urldate = {2025-01-08},
+	publisher = {bioRxiv},
[…]
