+++ docs/paper/paper.md
@@ -13,19 +13,27 @@ authors:
     affiliation: "1"
 
 affiliations:
- - name: "Université Paris Cité, Inserm U1133, CNRS UMR 8251, Paris, France"
+ - name: "Université Paris Cité, Inserm, CNRS, BFA, F-75013 Paris, France"
    index: 1
 
-date: 11 July 2024
+date: 8 January 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-The publication of AlphaFold 2[@jumper_highly_2021] has significantly advanced the field of protein structure prediction. The Prediction of protein structures has long constitued a central challenge in the field of structural bioinformatics, with the ultimate objective of elucidating the relationship between protein structure and function. The accurate prediction of protein structures is of great importance for a number of applications, including drug discovery, protein engineering, and the investigation of protein-protein interactions. AlphaFold, which employs a deep learning-based approach, has demonstrated unprecedented accuracy in protein structure prediction, outperforming other contemporary methods. In this paper, we present `af-analysis`, a Python package that provides tools for the analysis of Alphafold results. `af-analysis` has been designed to facilitate the analysis of protein structures predicted by Alphafold. It provides functions for the comparison of predicted structures with experimental structures, the visualisation of predicted structures, and the calculation of structural quality metrics.
+The publication of AlphaFold 2 [@jumper_highly_2021] has significantly advanced the field of protein structure prediction. The prediction of protein structures has long been a central challenge in the field of structural bioinformatics, with the ultimate goal of elucidating the relationship between protein structure and function. Accurate prediction of protein structure is essential for a number of applications, including drug discovery, protein engineering, and the study of protein-protein interactions. AlphaFold, which employs a deep learning-based approach, has demonstrated unprecedented accuracy in protein structure prediction, outperforming other contemporary methods. In this paper, we present `af-analysis`, a Python package that provides tools for the analysis of AlphaFold results. The `af-analysis` library has been designed to facilitate the analysis of many different protein structures predicted by AlphaFold and its derivatives. It provides functions for comparing predicted structures with experimental structures, visualising predicted structures, and calculating structural quality metrics.
 
 # Statement of need
 
[…]
