### 2025-01-08 #7577 @samuelmurail
Dear @AnjaConev ,

Thank you for your comments, feedback and testing.

> * Automated tests: pytest test_clustering.py is failing for me. I use Python version 3.11.7 and Rocky Linux 9.3 OS.
>   ![image](https://private-user-images.githubusercontent.com/7367905/400757225-ec7fceb7-51cf-4f2e-82ad-ce3e37ab9bc4.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzYzNDM3MDQsIm5iZiI6MTczNjM0MzQwNCwicGF0aCI6Ii83MzY3OTA1LzQwMDc1NzIyNS1lYzdmY2ViNy01MWNmLTRmMmUtODJhZC1jZTNlMzdhYjliYzQucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDEwOCUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAxMDhUMTMzNjQ0WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NGRjNmNlZDMzOTY5NzNmZDIzYzVjMTFmN2JhODRhNTdmYTdjZjEwNjRiMzNjNTZhZWM5Zjk0ODM4ZjBkNGRhYSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.EWirOG_aOH3QmzHjYLZfihDvBsm4IMriHWqNZnzsAlE)
>   clustering works when running it outside of the tests though.

I create a new conda env with python 3.11 but I could not reproduce your error. I suspect that `MDanalysis` or `sklearn` might be the cause of the problem as they are used only in the `clustering.py` module.
Could you reproduce the error and provide the full log of the error please ?

> * State of the field: You could add a paragraph to the paper about similar packages in the field and how yours is different.

I agree that the manuscript was a bit *light* on this part, so I added a paragraph on Alphafold analysis scores and existing tools. I could only find the AlphaBridge webserver and the PICKLUSTER plugin, if you are aware of additional tools, let me know.
