### 2025-01-09 #7577 @AnjaConev
Hi, it had something to do with my local setup of matplotlib - I fixed it and then all the tests worked!
Apologies for the confusion.

Great work! One final comment - it would be nice if you can add some instructions in the docs on how the package can be modified if we want to load models from a different modeling tool that you do not currently support.
