### 2025-01-09 #7577 @samuelmurail
> Hi, it had something to do with my local setup of matplotlib - I fixed it and then all the tests worked! Apologies for the confusion.

Great, thanks for the feedback.

> Great work! One final comment - it would be nice if you can add some instructions in the docs on how the package can be modified if we want to load models from a different modeling tool that you do not currently support.

Yes you are right, I was thinking to add something more flexible to import data. Here is my solution through python dictionnary, the procedure is explained in the "User Guide" -> "Importing procedure" :

Alternatively, you can use a python dictionary containing the `pdb` file list, the `.json` / `.npz` data file and the query name.

```python
  import af_analysis

  pdb_list = [
      "test/model_5_seed_002.pdb",
      "test/model_5_seed_007.pdb",
      "test/model_5_seed_003.pdb",
      "test/model_5_seed_006.pdb",
      "test/model_5_seed_004.pdb",
  ]
  json_list = [
      "test/model_5_seed_002.json",
      "test/model_5_seed_007.json",
      "test/model_5_seed_003.json",
      "test/model_5_seed_006.json",
      "test/model_5_seed_004.json",
  ]
  data_dict = {
      "data_file": json_list,
      "pdb": pdb_list,
      "query": "test_amyloid",
  }

  my_data = af_analysis.Data(data_dict=data_dict)
  # Extract the data that you know are present in the json/pickle files
  my_data.extract_fields(["plddt", "ptm", "iptm"])
```
