### 2025-01-09 646e141
Add possibility to load data throught dictionnary + black
