+++ src/af_analysis/format/chai1.py
@@ -34,14 +34,14 @@ def read_dir(directory):
 
     for file in os.listdir(pred_dir):
         if file.endswith(".cif"):
-            token = file[4: -4].split("_")[-1]
+            token = file[4:-4].split("_")[-1]
             model = int(token[-1])
 
             npz_dict = np.load(os.path.join(pred_dir, f"scores.model_idx_{model}.npz"))
 
-            #for key in npz_dict.keys():
+            # for key in npz_dict.keys():
             #    print(key, npz_dict[key])
-            #print('ratio', ( 0.2* npz_dict['ptm'] + 0.8*npz_dict['iptm'] )/npz_dict['aggregate_score'])
+            # print('ratio', ( 0.2* npz_dict['ptm'] + 0.8*npz_dict['iptm'] )/npz_dict['aggregate_score'])
 
             info_dict = {
                 "pdb": os.path.join(pred_dir, file),
@@ -63,19 +63,19 @@ def read_dir(directory):
             log_dict_list.append(info_dict)
 
     log_pd = pd.DataFrame(log_dict_list)
-    log_pd.loc[:,"query"] = os.path.basename(os.path.normpath(pred_dir))
+    log_pd.loc[:, "query"] = os.path.basename(os.path.normpath(pred_dir))
 
     # Update column names
     log_pd = log_pd.rename(
         columns={
             # https://github.com/jwohlwend/boltz/issues/73
-            #"confidence_score": "ranking_confidence",
+            # "confidence_score": "ranking_confidence",
             "ptm": "pTM",
             "iptm": "ipTM",
             "aggregate_score": "ranking_confidence",
         }
     )
-    
+
     # To ensure that tests are consistent across different systems
     # we sort the dataframe by pdb
     log_pd = log_pd.sort_values(by=["pdb"]).reset_index(drop=True)
