+++ src/af_analysis/clustering.py
@@ -18,7 +18,7 @@ __author__ = "Alaa Reguei, Samuel Murail"
 __copyright__ = "Copyright 2023, RPBS"
 __credits__ = ["Samuel Murail", "Alaa Reguei"]
 __license__ = "GNU General Public License version 2"
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 __maintainer__ = "Samuel Murail"
 __email__ = "samuel.murail@u-paris.fr"
 __status__ = "Beta"
