+++ src/af_analysis/analysis.py
@@ -10,7 +10,7 @@ __author__ = "Alaa Reguei"
 __copyright__ = "Copyright 2023, RPBS"
 __credits__ = ["Samuel Murail", "Alaa Reguei"]
 __license__ = "GNU General Public License version 2"
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 __maintainer__ = "Samuel Murail"
 __email__ = "samuel.murail@u-paris.fr"
 __status__ = "Beta"
