+++ README.md
@@ -7,22 +7,21 @@
 [![License: GPL v2](https://img.shields.io/badge/License-GPL%20v2-blue.svg)](https://www.gnu.org/licenses/old-licenses/gpl-2.0.html)
 [![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/samuelmurail/af_analysis/blob/main/basic_example_colab.ipynb)
 
-
 # About Alphafold Analysis
 
 <img src="https://raw.githubusercontent.com/samuelmurail/af_analysis/master/docs/source/logo.jpeg" alt="AF Analysis Logo" width="300" style="display: block; margin: auto;"/>
 
-
 `af-analysis` is a python package for the analysis of AlphaFold protein structure predictions.
 This package is designed to simplify and streamline the process of working with protein structures
 generated by:
+
 * [AlphaFold 2][AF2]
 * [AlphaFold 3][AF3]
 * [ColabFold][ColabFold]
 * [AlphaFold-Multimer][AF2-M]
 * [AlphaPulldown][AlphaPulldown]
-* [Boltz1][Boltz1].
-* [Chai-1][Chai1].
+* [Boltz1][Boltz1]
+* [Chai-1][Chai1]
 
 
 Source code repository:
@@ -35,13 +34,13 @@ Analyzing the abundance of resulting structural models can be challenging and ti
 Existing tools often require separate scripts for calculating various quality metrics (pDockQ, pDockQ2, LIS score) and assessing model diversity.
 `af-analysis` addresses these challenges by providing a unified and user-friendly framework for in-depth analysis of AlphaFold 2 results.
 
-## Main features:
+## Main features
 
 * Import AlphaFold or ColabFold prediction directories as pandas DataFrames for efficient data handling.
 * Calculate and add additional structural quality metrics to the DataFrame, including:
-    * pDockQ
-    * pDockQ2
-    * LIS score
+  * pDockQ
+  * pDockQ2
+  * LIS score
 * Visualize predicted protein models.
 * Cluster generated models to identify diverse conformations.
 * Select the best models based on defined criteria.
@@ -49,19 +48,19 @@ Existing tools often require separate scripts for calculating various quality me
 
 ## Installation
 
-- `af-analysis` is available on PyPI and can be installed using ``pip``:
+* `af-analysis` is available on PyPI and can be installed using ``pip``:
 
 ```bash
 pip install af_analysis
 ```
 
-- You can install last version from the github repo:
+* You can install last version from the github repo:
 
 ```bash
 pip install git+https://github.com/samuelmurail/af_analysis.git@main
 ```
 
-- AF-Analysis can also be installed easily through github:
+* AF-Analysis can also be installed easily through github:
 
 ```bash
 git clone https://github.com/samuelmurail/af_analysis
@@ -81,7 +80,6 @@ The complete documentation is available at [ReadTheDocs](https://af-analysis.rea
 
 ## Usage
 
-
 ### Importing data
 
 Create the `Data` object, giving the path of the directory containing the results of the alphafold2/colabfold run. 
@@ -99,7 +97,7 @@ my_data.df
 
 ### Analysis
 
-- The `analysis` package contains several function to add metrics like [pdockQ][pdockq] and [pdockQ2][pdockq2]:
[…]
