+++ docs/paper/paper.md
@@ -7,10 +7,11 @@ tags:
   - Structural bioinformatics
 authors:
   - name: Alaa Reguei
-    affiliation: "1"
+    affiliation: 1
   - name: Samuel Murail
     orcid: "0000-0002-3842-5333"
-    affiliation: ["1", "2"]
+    corresponding: true
+    affiliation: [1, 2]
 
 affiliations:
  - name: "Université Paris Cité, Inserm, CNRS, BFA, F-75013 Paris, France"
