+++ docs/paper/paper.md
@@ -11,7 +11,7 @@ authors:
   - name: Samuel Murail
     orcid: "0000-0002-3842-5333"
     corresponding: true
-    affiliation: [1, 2]
+    affiliation: "[1, 2]"
 
 affiliations:
  - name: "Université Paris Cité, Inserm, CNRS, BFA, F-75013 Paris, France"
