+++ docs/paper/paper.md
@@ -11,12 +11,12 @@ authors:
   - name: Samuel Murail
     orcid: "0000-0002-3842-5333"
     corresponding: true
-    affiliation: "[1, 2]"
+    affiliation: "1, 2"
 
 affiliations:
- - name: "Université Paris Cité, Inserm, CNRS, BFA, F-75013 Paris, France"
+ - name: Université Paris Cité, Inserm, CNRS, BFA, F-75013 Paris, France
    index: 1
- - name: "Ressource Parisienne en Bioinformatique Structurale (RPBS), F-75013 Paris, France."
+ - name: Ressource Parisienne en Bioinformatique Structurale (RPBS), F-75013 Paris, France
    index: 2
 
 date: 8 January 2025
