### 2025-01-13 #7577 @samuelmurail
Dear @srmnitc, @AnjaConev and @rvhonorato,

First of all, I'd like to thank @AnjaConev and @rvhonorato for their constructive reviews.

Thanks to @srmnitc for the editorial work, I think that both reviewers are happy with the current version of the library.
Since the submission of the paper, the library has changed name (af2_analysis -> af_analysis) and version (v0.0.2 to v 0.1.3).

Do I need to do anything on my end to update the name and version of the package ?

Cheers,
Samuel
