### 2025-01-30 #7577 @srmnitc
@samuelmurail Thanks for making all the changes, and @AnjaConev and @rvhonorato for the reviews. Unfortunately since I was on mostly away until now, this took a bit of time to get around to.

@AnjaConev and @rvhonorato could you please confirm that you recommend the acceptance of this work in JOSS? After that I can go ahead with the rest of the steps.
