### 2025-02-04 231acb2
Fix a bug with show_info() add analysis of PAE matrix
