+++ src/af_analysis/__init__.py
@@ -8,7 +8,7 @@ __author__ = "Samuel Murail"
 __copyright__ = "Copyright 2023, RPBS"
 __credits__ = ["Samuel Murail"]
 __license__ = "GNU General Public License version 2"
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 __maintainer__ = "Samuel Murail"
 __email__ = "samuel.murail@u-paris.fr"
 __status__ = "Beta"
