### 2025-02-12 #7577 @samuelmurail
@srmnitc Thanks for the feedback.

> * Double check authors and affiliations (including ORCIDs)
* Done
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
* The version is 0.1.4 (github release, zenodo and pypi)
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
* Here is the doi:  [10.5281/zenodo.14859764](https://zenodo.org/records/14859764)
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
* Yes the titles on JOSS paper and Zenodo are matching, same as for author list.
**However** note that the papaer title as changed since the submission from `Af2-analysis: a Python package for Alphafold 2 analysis` to `Af-analysis: a Python package for Alphafold analysis`
> * Make sure that the license listed for the archive is the same as the software license.
* Same license for zenodo and github (GNU General Public License v2.0).
