### 2025-02-17 samuelmurail/af_analysis#5 @srmnitc
Changes to the JOSS paper
This is part of the review over at https://github.com/openjournals/joss-reviews/issues/7577

Here are some small changes to the paper from my side:

- Does Alaa Reguei have an ORCID?
- line 9: could be strengthened with a citation
- line 42-43: `Furthermore, if the quality metrics produced by AlphaFold are good,
43 additional metrics have been developed to assess the quality of the models.`, would  `Furthermore, while the quality metrics produced by AlphaFold are good,
43 additional metrics have been developed to assess the quality of the models.` make better sense? Please take a look.
- references: please use either journal abbreviations or the full names consistently

Please feel free to close this issue when the points are addressed. We can finish up the paper after this. Thanks again for the nice package!
