+++ docs/paper/paper.md
@@ -7,6 +7,7 @@ tags:
   - Structural bioinformatics
 authors:
   - name: Alaa Reguei
+  - orcid: "0009-0006-1718-4864"
     affiliation: 1
   - name: Samuel Murail
     orcid: "0000-0002-3842-5333"
@@ -19,13 +20,13 @@ affiliations:
  - name: Ressource Parisienne en Bioinformatique Structurale (RPBS), F-75013 Paris, France
    index: 2
 
-date: 8 January 2025
+date: 17 February 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-The publication of AlphaFold 2 [@jumper_highly_2021] has significantly advanced the field of protein structure prediction. The prediction of protein structures has long been a central challenge in the field of structural bioinformatics, with the ultimate goal of elucidating the relationship between protein structure and function. Accurate prediction of protein structure is essential for a number of applications, including drug discovery, protein engineering, and the study of protein-protein interactions. AlphaFold, which employs a deep learning-based approach, has demonstrated unprecedented accuracy in protein structure prediction, outperforming other contemporary methods. In this paper, we present `af-analysis`, a Python package that provides tools for the analysis of AlphaFold results. The `af-analysis` library has been designed to facilitate the analysis of many different protein structures predicted by AlphaFold and its derivatives. It provides functions for comparing predicted structures with experimental structures, visualising predicted structures, and calculating structural quality metrics.
+The publication of AlphaFold 2 [@jumper_highly_2021] has significantly advanced the field of protein structure prediction. The prediction of protein structures has long been a central challenge in the field of structural bioinformatics, with the ultimate goal of elucidating the relationship between protein structure and function [@baker_protein_2001; @pearce_toward_2021]. Accurate prediction of protein structure is essential for a number of applications, including drug discovery, protein engineering, and the study of protein-protein interactions. AlphaFold, which employs a deep learning-based approach, has demonstrated unprecedented accuracy in protein structure prediction, outperforming other contemporary methods. In this paper, we present `af-analysis`, a Python package that provides tools for the analysis of AlphaFold results. The `af-analysis` library has been designed to facilitate the analysis of many different protein structures predicted by AlphaFold and its derivatives. It provides functions for comparing predicted structures with experimental structures, visualising predicted structures, and calculating structural quality metrics.
 
 # Statement of need
 
@@ -35,8 +36,7 @@ Analysis of AlphaFold results is a crucial step in the process of utilising thes
 
 Although these tools are very practical, Bjorn Wallner has shown that calculating 5 or 25 basic AlphaFold models may not be enough, it is sometimes necessary to generate thousands of models to obtain a few high quality models, leading to the AlphaFold derivative, AFsample [@10.1093/bioinformatics/btad573]. Massive sampling altogether with multiple software usage (AFsample and ColabFold), weights and parameters has been integrated into the MassiveFold software [@raouraoua_massivefold_2024] and has shown performance approaching the accuracy of AlphaFold 3.
 
-
 The subsequent analysis of hundreds to thousands of models can prove to be a tedious and meticulous process, as dealing with thousands of models and different output formats can be time consuming.
[…]
