+++ docs/paper/paper.bib
@@ -27,7 +27,7 @@
   pages={679--682},
   year={2022},
   doi = {10.1038/s41592-022-01488-1},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
 }
 
 @article {Evans2021.10.04.463034,
@@ -36,11 +36,11 @@
 	elocation-id = {2021.10.04.463034},
 	year = {2022},
 	doi = {10.1101/2021.10.04.463034},
-	publisher = {Cold Spring Harbor Laboratory},
 	abstract = {While the vast majority of well-structured single protein chains can now be predicted to high accuracy due to the recent AlphaFold [1] model, the prediction of multi-chain protein complexes remains a challenge in many cases. In this work, we demonstrate that an AlphaFold model trained specifically for multimeric inputs of known stoichiometry, which we call AlphaFold-Multimer, significantly increases accuracy of predicted multimeric interfaces over input-adapted single-chain AlphaFold while maintaining high intra-chain accuracy. On a benchmark dataset of 17 heterodimer proteins without templates (introduced in [2]) we achieve at least medium accuracy (DockQ [3] >= 0.49) on 13 targets and high accuracy (DockQ >= 0.8) on 7 targets, compared to 9 targets of at least medium accuracy and 4 of high accuracy for the previous state of the art system (an AlphaFold-based system from [2]). We also predict structures for a large dataset of 4,446 recent protein complexes, from which we score all non-redundant interfaces with low template identity. For heteromeric interfaces we successfully predict the interface (DockQ >= 0.23) in 70\% of cases, and produce high accuracy predictions (DockQ >= 0.8) in 26\% of cases, an improvement of +27 and +14 percentage points over the flexible linker modification of AlphaFold [4] respectively. For homomeric inter-faces we successfully predict the interface in 72\% of cases, and produce high accuracy predictions in 36\% of cases, an improvement of +8 and +7 percentage points respectively.Competing Interest StatementThe authors have declared no competing interest.},
 	URL = {https://www.biorxiv.org/content/early/2022/03/10/2021.10.04.463034},
 	eprint = {https://www.biorxiv.org/content/early/2022/03/10/2021.10.04.463034.full.pdf},
-	journal = {bioRxiv}
+	journal = {bioRxiv},
+	publisher = {Cold Spring Harbor Laboratory},
 }
 
 @article{10.1093/bioinformatics/btad573,
@@ -75,19 +75,6 @@
     eprint = {https://academic.oup.com/bioinformatics/article-pdf/39/7/btad424/50886203/btad424.pdf},
 }
 
-@article {Kim2024.02.19.580970,
-	author = {Kim, Ah-Ram and Hu, Yanhui and Comjean, Aram and Rodiger, Jonathan and Mohr, Stephanie E. and Perrimon, Norbert},
-	title = {Enhanced Protein-Protein Interaction Discovery via AlphaFold-Multimer},
-	elocation-id = {2024.02.19.580970},
-	year = {2024},
-	doi = {10.1101/2024.02.19.580970},
-	publisher = {Cold Spring Harbor Laboratory},
[…]
