### 2025-02-17 samuelmurail/af_analysis#5 @samuelmurail
> * Does Alaa Reguei have an ORCID?
* Alaa's ORCID has been added (I asked him to create one).
> * line 9: could be strengthened with a citation
* I have added two citations `(Baker & Sali, 2001; Pearce & Zhang, 2021)`
> * line 42-43: `Furthermore, if the quality metrics produced by AlphaFold are good, 43 additional metrics have been developed to assess the quality of the models.`, would  `Furthermore, while the quality metrics produced by AlphaFold are good, 43 additional metrics have been developed to assess the quality of the models.` make better sense? Please take a look.
* Done, thanks for the suggestion
>  * references: please use either journal abbreviations or the full names consistently
* Done, I also noticed that some bioarxiv citation had not the `articlr` format, this is corrected

> Please feel free to close this issue when the points are addressed. We can finish up the paper after this. Thanks again for the nice package!

Thank you so much for editing this paper, I really enjoy the whole JOSS article submission process and its transparency.

Cheers,
Samuel
