### 2025-02-17 dc70909
Fix issue with missing affiliation
