### 2025-02-17 #7577 @samuelmurail
@srmnitc issue with pdf generation is fixed.
