### 2025-02-17 #7577 @srmnitc
> [@srmnitc](https://github.com/srmnitc) issue with pdf generation is fixed.

that was quick, thanks!
