### 2025-02-17 #7577 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
