### 2025-02-20 #7577 @Kevin-Mattheus-Moerman
- [ ] @samuelmurail `visualising and utilising` are British English while `characterize` is American English, I suggest you choose one style and use it consistently.
