### 2025-02-22 47abb6f
Remove american english word
