### 2025-02-22 #7577 @samuelmurail
> * [ ]  [@samuelmurail](https://github.com/samuelmurail) `visualising and utilising` are British English while `characterize` is American English, I suggest you choose one style and use it consistently.

Done I have switched to British English, I also switched an `analyze` to `analyse` .
