+++ CITATION.cff
@@ -1,0 +1,36 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Reguei
+  given-names: Alaa
+  orcid: "https://orcid.org/0009-0006-1718-4864"
+- family-names: Murail
+  given-names: Samuel
+  orcid: "https://orcid.org/0000-0002-3842-5333"
+contact:
+- family-names: Murail
+  given-names: Samuel
+  orcid: "https://orcid.org/0000-0002-3842-5333"
+doi: 10.5281/zenodo.14859764
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Reguei
+    given-names: Alaa
+    orcid: "https://orcid.org/0009-0006-1718-4864"
+  - family-names: Murail
+    given-names: Samuel
+    orcid: "https://orcid.org/0000-0002-3842-5333"
+  date-published: 2025-03-14
+  doi: 10.21105/joss.07577
+  issn: 2475-9066
+  issue: 107
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7577
+  title: "Af-analysis: a Python package for Alphafold analysis"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07577"
+  volume: 10
+title: "Af-analysis: a Python package for Alphafold analysis"
