### 2025-03-17 #7577 @samuelmurail
> [@samuelmurail](https://github.com/samuelmurail) congratulations on this JOSS publication!
>
> Thanks for editing [@srmnitc](https://github.com/srmnitc) !!
>
> And a special thank you to the reviewers: [@AnjaConev](https://github.com/AnjaConev), [@rvhonorato](https://github.com/rvhonorato) !!!

Thanks a lot for all the editorial work @Kevin-Mattheus-Moerman and @srmnitc , and thanks for the reviewers @AnjaConev and @rvhonorato.
Once again I really appreciate the submission process at JOSS.
